Reduced: Lines of Codes
